+++ src/utils/warnings/warnings.rs
@@ -1,0 +1,19 @@
+#[path = "../ui/color_texts.rs"]
+mod font_styles;
+
+pub enum Warnings {
+    NoArgsWarning,
+}
+
+impl Warnings {
+    // Function which throws a warning
+    pub fn throw_warning(&self) {
+        match *self {
+            Warnings::NoArgsWarning => font_styles::FontStyle::_warning_message(
+                &font_styles::FontStyle::Orange,
+                String::from("NoArgsWarning"),
+                String::from("No arguments were passed."),
+            ),
+        }
+    }
+}